# Optimize the generated site