--- a/.scalafmt.conf
+++ b/.scalafmt.conf
@@ -3,4 +3,4 @@ align = none
 danglingParentheses = true
 newlines.neverBeforeJsNative = true
 newlines.sometimesBeforeColonInMethodReturnType = false
-
+assumeStandardLibraryStripMargin = true