--- a/metadoc-core/src/main/protobuf/metadoc.proto
+++ b/metadoc-core/src/main/protobuf/metadoc.proto
@@ -10,6 +10,10 @@ message Position {
     int32 endCharacter = 5;
 }
 
+message SymbolIndexes {
+    repeated SymbolIndex indexes = 1;
+}
+
 message SymbolIndex {
     string symbol = 1;
     Position definition = 2;