--- a/metadoc-core/src/main/scala/metadoc/MetadocEnrichments.scala
+++ b/metadoc-core/src/main/scala/metadoc/MetadocEnrichments.scala
@@ -2,6 +2,7 @@ package metadoc
 
 import metadoc.{schema => d}
 import scala.meta.internal.{semanticdb => s}
+import scala.meta.internal.semanticdb.Scala._
 
 object MetadocEnrichments {
   implicit class XtensionMetadocRange(val r: d.Range) extends AnyVal {
@@ -14,6 +15,13 @@ object MetadocEnrichments {
         r.endCharacter
       )
     }
+  }
+  implicit class XtensionSymbolString(val symbol: String) extends AnyVal {
+    def symbolIndexPath: String = toplevelPackage + "/package.symbolindexes.gz"
+    def toplevelPackage: String =
+      if (symbol.isNone) symbol
+      else if (symbol.isPackage) symbol
+      else symbol.owner.toplevelPackage
   }
   implicit class XtensionSemanticdbRange(val r: s.Range) extends AnyVal {
     def toDocRange: d.Range = {