--- a/metadoc-tests/src/test/scala/metadoc/tests/GeneratedSiteEnrichments.scala
+++ b/metadoc-tests/src/test/scala/metadoc/tests/GeneratedSiteEnrichments.scala
@@ -1,0 +1,24 @@
+package metadoc.tests
+import java.nio.file.Files
+import java.util.zip.GZIPInputStream
+import scala.meta.io.AbsolutePath
+import scalapb.GeneratedMessage
+import scalapb.GeneratedMessageCompanion
+import scalapb.Message
+
+object GeneratedSiteEnrichments {
+  implicit class XTensionGeneratedMessageCompanion[
+      A <: GeneratedMessage with Message[A]
+  ](companion: GeneratedMessageCompanion[A]) {
+    def parseFromCompressedPath(path: AbsolutePath): A = {
+      val in = Files.newInputStream(path.toNIO)
+      val gin = new GZIPInputStream(in)
+      try companion.parseFrom(gin)
+      finally {
+        gin.close()
+        in.close()
+      }
+    }
+  }
+
+}