… metadoc-js/src/main/scala/metadoc/JSSha512.scala
… metadoc-js/yarn.lock