This commit implements two optimizations to make the generate site
contain fewer files and be smaller in bytes.

- gzip all generated files: this required changing the JavaScript client
  to inflate responses with the library pako (http://nodeca.github.io/pako/#inflate).
  I originally hoped the browser would automatically inflate the
  reponses us by using the `*.gz` file extension. The response headers
  for the webpack + browsersync + GitHub pages all return
  `Content-Encoding: gzip` for filenames with the `*.gz` extension.
  Nevertheless, the fetch response body contained deflated bytes
  so I resorted to inflating in JavaScript.
- group symbol indexes by package. This primarily reduces the number of
  generated files. Most symbol indexes were only a few bytes but
  occupied the minimum block-size on disk which is 4kb on my macbook.
  For a Scalameta + Scalafix combined metadoc site, this change
  made the total symbol index go from taking 40mb on disk to 3.4mb!

To measure these optimizations I ran metadoc against three different
corpora.

SemanticDB index:

| Corpus               | Before | After |
| -------------------- | ------ | ----- |
| paiges               | 332kb  | 112kb |
| scalameta + scalafix | 24mb   | 5.9mb |
| megasite             | 1.25gb | 209mb |

Symbol index:

| Corpus               | Before | After |
| -------------------- | ------ | ----- |
| paiges               | 776kb  | 16kb  |
| scalameta + scalafix | 45mb   | 888kb |
| megasite             | n/a    | 44mb  |

The "megasite" corpus is a combination of 55 projects with a
total of 3.5 million lines of code. Generating a metadoc site for the
megasite takes ~50 seconds on my laptop.