@olafurpg 2018-09-27
Grouping by top-level package is problematic for widely referenced packages, for example the  `org.scalatest` symbol index in the megasite is 11mb. I was thinking of chunking down those files in ~500kb sizes and the client needs to "guess" the first package index and from there we can link to a secondary package index. That would mean users you never download more than ~1mb to resolve a symbol (and we can lower 1mb at the cost of producing more files).

Either way, I believe even 11mb is an acceptable solution considering the megasite is for a corpus of 3.5 million loc corpus. We can further improve the situation in followup PRs.