@olafurpg 2018-09-27
Here is a live demo with the combined scalameta + scalafix site: https://olafurpg.github.io/metadoc/#/scalameta/scalameta/io/shared/src/main/scala/scala/meta/io/AbsolutePath.scala#L14C28-L14C40

Try "Find all references" on `AbsolutePath`