@olafurpg 2018-09-28
Rebased on master and updated the demo to include scala-library https://olafurpg.github.io/metadoc/#/scala/Predef.scala#L134C8-L134C14