@olafurpg 2018-09-28
CI went from 10 minutes to 5 minutes for scripted tests on sbt 1 after rebasing on top of #101, awesome work @jonas