@olafurpg 2018-09-28
No more CI timeouts \o/

Guess not having to downloads 350mb of icons helps 😂